requirements.txt
torch==2.7.1
torchvision==0.22.1
transformers==4.53.1
numpy==1.26.4
Pillow>=8.0.0
pydantic>=2.0.0
timm>=0.9.0
tqdm>=4.60.0
pyzmq>=25.0.0
transforms3d>=0.4.0
safetensors>=0.3.0
typing-extensions>=4.0.0
opencv-python
matplotlib
termcolor
einops
